+++ Draw_lib/draw.h
@@ -4,14 +4,14 @@
 #include "../mzapo_lib/mzapo.h"
 
 typedef struct {
-    uint16_t x, y;
+    int x, y;
 } Vertex_2D;
 
 // Basic 2D shapes
 void draw_rectangle(lcdpixel* fb, Vertex_2D topLeft, Vertex_2D bottomRight, lcdpixel color);
 void draw_line(lcdpixel* fb, Vertex_2D start, Vertex_2D end, lcdpixel color);
 void draw_ellipse(lcdpixel* fb, Vertex_2D center, int a, int b, lcdpixel color);
-void draw_circle(lcdpixel* fb, Vertex_2D center, int radius, lcdpixel color);
+void draw_circle(lcdpixel* fb, Vertex_2D center, float radius, lcdpixel color);
 void clear_display(lcdpixel* fb);
 
 #endif
